#845 move osmaxx.conversion_api to osmaxx.conversion.constants
