+++ osmaxx/excerptexport/forms/order_options_mixin.py
@@ -3,7 +3,7 @@ from django import forms
 from django.utils.translation import ugettext_lazy as _
 
 from osmaxx.conversion.converters.converter_gis.detail_levels import DETAIL_LEVEL_CHOICES
-from osmaxx.conversion_api import formats, coordinate_reference_systems as crs
+from osmaxx.conversion.constants import formats, coordinate_reference_systems as crs
 
 
 class OrderOptionsMixin(forms.Form):
