+++ tests/excerptexport/test_conversion_api_client.py
@@ -9,8 +9,8 @@ from requests import HTTPError
 from rest_framework.reverse import reverse
 
 from osmaxx.api_client import ConversionApiClient, API_client
-from osmaxx.conversion_api.formats import FGDB, SPATIALITE
-from osmaxx.conversion_api.statuses import RECEIVED
+from osmaxx.conversion.constants.formats import FGDB, SPATIALITE
+from osmaxx.conversion.constants.statuses import RECEIVED
 from osmaxx.excerptexport.models import Excerpt, ExtractionOrder
 from osmaxx.job_progress.views import tracker
 from tests.test_helpers import vcr_explicit_path as vcr
