+++ tests/excerptexport/test_models.py
@@ -3,8 +3,8 @@ from django.contrib.auth.models import User
 from django.contrib.gis import geos
 from hamcrest import assert_that, contains_inanyorder as contains_in_any_order
 
-from osmaxx.conversion_api.formats import GARMIN
-from osmaxx.conversion_api.statuses import FINISHED
+from osmaxx.conversion.constants.formats import GARMIN
+from osmaxx.conversion.constants.statuses import FINISHED
 from osmaxx.excerptexport import models
